requirements.txt
flask==2.3.3
flask-cors==4.0.0
tensorflow>=2.15.0
numpy>=1.24.3
Pillow>=10.0.1
opencv-python>=4.8.1
requests>=2.28.0
urllib3>=1.26.0

